Reorganizing into packages. Much nicer organization now.
